backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
requests==2.32.5
geopy==2.4.1
python-dateutil==2.9.0.post0
django-cors-headers==4.9.0

